requirements.txt
pywebview
pathlib
typing
send2trash
